pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = ""
authors = ["Stanislav Kulchitskij <kulchitskihstas42@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.8"}
beanie = "^1.29.0"
pydantic = "^2.10.6"
passlib = "^1.7.4"
pyjwt = "^2.10.1"
pytest = "^8.3.4"
python-multipart = "^0.0.20"
pytest-asyncio = "^0.25.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
